Feat: 사용자 얼굴 이미지 S3로 로드
