+++ src/views/Camera/index.tsx
@@ -10,7 +10,8 @@ import * as types from '../../types/cam';
 import constraints from '../../common/constraints';
 import uploadToS3Bucket from '../../services/Cam/uploadToS3Bucket';
 import EmotionSetData from './EmotionSetData';
-import loadUsim from '../../services/Cam/loadUsim';
+import config from '../../config';
+import axios from 'axios';
 import Sparky from 'react-sparkle';
 import { ConnectingAirportsOutlined } from '@mui/icons-material';
 
@@ -128,11 +129,29 @@ function CameraPage(props: any) {
     // 라벨링 할 인물 이미지 로컬에서 가져오기
     const loadImage = async () => {
         // 업로드 된 이미지 이름을 배열에 담아 라벨링 합니다.
-        const labels = [`${userId}`];
+        const resultUrls: any[] = [];
+        let labels = await axios({
+            url: `http://${config.server.host}:${config.server.port}/camera/usim/${userId}`,
+            method: 'get',
+        })
+            .then(function (result) {
+                result.data.forEach((element: any) => {
+                    console.log(element);
+                    resultUrls.push(element.userImgUrl);
+                });
+                console.log(resultUrls);
+                console.log('이미지 로드 성공');
+                return resultUrls;
+            })
+            .catch(function (error) {
+                console.log(error);
+                console.log('이미지 로드 실패');
+                return resultUrls;
+            });
 
         return Promise.all(
             labels.map(async (label) => {
-                const images = await faceapi.fetchImage(require(`./imgs/${label}.jpg`));
+                const images = await faceapi.fetchImage(label);
                 const descriptions = [];
                 const detections = await faceapi.detectSingleFace(images).withFaceLandmarks().withFaceDescriptor();
                 if (detections) descriptions.push(detections.descriptor);
@@ -196,6 +215,7 @@ function CameraPage(props: any) {
                     const drawBox = new faceapi.draw.DrawBox(box, { boxColor: 'red', label: label });
 
                     if (label === userId) drawBox.draw(canvas); // 특정 사용자가 감지됐을 때만 바운딩 박스 표시
+                    // drawBox.draw(canvas); // 전체 사용자 인식
 
                     // faceapi.draw.drawFaceExpressions(canvas, resizedDetections); // 감정 수치 표시
